// types/tasks.ts
// Represents a task.
export interface ITask {
    // The ID of the task.
    id: string;
  
    // The title of the task.
    title: string;
  
    // The description of the task.
    description: string;
  
    // The status of the task (e.g., "To Do," "In Progress," "Completed").
    status: string;
  }
  